JSON files specify their own $schema
VSCode understands the "$schema" property of a JSON object: https://code.visualstudio.com/docs/languages/json#_mapping-in-the-json

check-jsonschema has an open feature request to support this which looks positive: https://github.com/python-jsonschema/check-jsonschema/issues/310

This commit adds a "$schema" property to the top level of the JSON files.
It also moves the schema to a separate directory to clean up the GitHub workflow.